Split ffi.rs into focused sub-modules (#601)
Split the monolithic src/ffi.rs (1,075 lines) into focused sub-modules
under src/ffi/, grouped by FFI concern:

- ffi/mod.rs — re-exports, free_discovery_result
- ffi/gpu.rs — check_gpu_available
- ffi/recording.rs — record_discovery, streaming session functions
- ffi/analysis.rs — rank_focus_neurons, analyze_parallel
- ffi/utilities.rs — merge, read, export, version

All 13 #[no_mangle] FFI symbols remain visible in the compiled dylib.
