+++ AGENTS.md
@@ -38,7 +38,12 @@ admitted back into the population.
 ```
 src/
 ├── lib.rs                    # Module declarations, re-exports, version init
-├── ffi.rs                    # FFI entry points (no_mangle extern "C")
+├── ffi/                      # FFI entry points (no_mangle extern "C") (Issue #601)
+│   ├── mod.rs                # Public API, re-exports, free_discovery_result
+│   ├── gpu.rs                # GPU probe entry points (check_gpu_available)
+│   ├── recording.rs          # Recording entry points (streaming and single-call)
+│   ├── analysis.rs           # Analysis entry points (rank_focus_neurons, analyze_parallel)
+│   └── utilities.rs          # Utility entry points (merge, read, export, version)
 ├── ffi_types/                # JSON request/response structs for FFI boundary (Issue #596)
 │   ├── mod.rs                # Public API, re-exports, shared types (Creature, Neuron, Synapse)
 │   ├── requests.rs           # FFI request structs (input from NEAT-AI)
